Extracted validation functions to lib/utils/formValidation
