+++ data-dashboard/pages/signup.tsx
@@ -6,6 +6,7 @@ import type { MouseEvent } from 'react';
 import { IbmButton, TextInput, InputField } from '@/lib/components';
 import trpc from '@/lib/hooks/trpc';
 import TopLayout from '@/lib/components/TopLayout/TopLayout';
+import { validateEmail, validatePassword } from '@/lib/utils/formValidation';
 import { NextPageWithLayout } from './_app';
 
 const Body = styled.main`
@@ -106,50 +107,6 @@ function SignUp() {
     confirmPassword: '',
   });
 
-  const validatePassword = (password: string) => {
-    const passwordRequirements = {
-      min: 8,
-      max: 24,
-      upper: 1,
-      lower: 1,
-      number: 1,
-      special: 1,
-    };
-    // Check password requirements one by one
-    if (password.length < passwordRequirements.min) {
-      return `La contraseña debe tener al menos ${passwordRequirements.min} caracteres`;
-    }
-    if (password.length > passwordRequirements.max) {
-      return `La contraseña debe tener menos de ${passwordRequirements.max} caracteres`;
-    }
-    if (password.length - password.replace(/[A-Z]/g, '').length < passwordRequirements.upper) {
-      return `La contraseña debe tener al menos ${passwordRequirements.upper} mayúscula`;
-    }
-    if (password.length - password.replace(/[a-z]/g, '').length < passwordRequirements.lower) {
-      return `La contraseña debe tener al menos ${passwordRequirements.lower} minúscula`;
-    }
-    if (password.length - password.replace(/[0-9]/g, '').length < passwordRequirements.number) {
-      return `La contraseña debe tener al menos ${passwordRequirements.number} número`;
-    }
-    if (password.length - password.replace(/[^a-zA-Z0-9]/g, '').length < passwordRequirements.special) {
-      return `La contraseña debe tener al menos ${passwordRequirements.special} caracter especial`;
-    }
-    return '';
-  };
-
-  const validateEmail = (email: string) => {
-    const emailRegex = new RegExp(
-      // eslint-disable-next-line no-control-regex
-      '^[a-zA-Z0-9.!#$%&\'*+/=?^_`{|}~-]+@[a-zA-Z0-9]'
-      + '(?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?'
-      + '(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$',
-    );
-    if (!emailRegex.test(email)) {
-      return 'Ingresa un correo electrónico válido';
-    }
-    return '';
-  };
-
   const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
     const { name, value } = event.target;
     setCredentials((prevCredentials) => ({
@@ -217,7 +174,7 @@ function SignUp() {
                 }}
               />
               {
-                error.email && <Error className='error-message'>{error.email}</Error>
+                error.email && <Error className="error-message">{error.email}</Error>
               }
               <InputField
                 label="Contraseña"
@@ -233,7 +190,7 @@ function SignUp() {
                 }}
               />
               {
-                error.password && <Error className='error-message'>{error.password}</Error>
+                error.password && <Error className="error-message">{error.password}</Error>
               }
               <InputField
                 label="Confirmar contraseña"
@@ -249,7 +206,7 @@ function SignUp() {
                 }}
               />
               {
-                error.confirmPassword && <Error className='error-message'>{error.confirmPassword}</Error>
+                error.confirmPassword && <Error className="error-message">{error.confirmPassword}</Error>
               }
             </InputFields>
             <IbmButton
